Make replaceElement function part of the public API
