+++ src/index.js
@@ -1,5 +1,6 @@
 import icons from './icons';
 import toSvg from './to-svg';
 import replace from './replace';
+import replaceElement from './replace-element';
 
-module.exports = { icons, toSvg, replace };
+module.exports = { icons, toSvg, replace, replaceElement };
